# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Web-Karir (Career Hub) Secure REST API",
        "description": "Dokumentasi REST API lengkap dan terproteksi untuk platform Web-Karir (Career Hub). Menggunakan Laravel Sanctum Bearer Token Authentication, Rate Limiting, dan Spatie Role-Based Access Control.",
        "contact": {
            "name": "Web-Karir Support Team",
            "email": "admin@webkarir.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Web-Karir Primary API Server"
        }
    ],
    "paths": {
...
        "/admin/users": {
            "get": {
                "tags": [
                    "Super Admin Moderation"
                ],
                "summary": "[Super Admin] Daftar Seluruh Pengguna Platform",
                "description": "Mengambil seluruh daftar pengguna (Candidate, HR, Super Admin).",
                "operationId": "4da9f9b8d631a791611f4799c05e36ae",
                "responses": {
                    "200": {
                        "description": "Daftar pengguna",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Daftar pengguna berhasil diambil.",
                                    "data": {
                                        "current_page": 1,
                                        "data": [
                                            {
                                                "id": 1,
                                                "name": "Budi Pratama",
                                                "email": "budi@example.com",
                                                "is_suspended": false,
                                                "roles": [
                                                    {
                                                        "name": "Candidate"
                                                    }
                                                ]
                                            }
                                        ]
                                    },
                                    "errors": null
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{id}/toggle-suspend": {
            "patch": {
                "tags": [
                    "Super Admin Moderation"
                ],
                "summary": "[Super Admin] Suspend / Unsuspend Akun User",
                "description": "Menangguhkan atau mengaktifkan kembali akun pengguna.",
                "operationId": "4b65d737773f2d1b03660d9e7bc030b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status penangguhan akun diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Akun pengguna Budi Pratama berhasil ditangguhkan (suspended).",
                                    "data": {
                                        "id": 1,
                                        "name": "Budi Pratama",
                                        "is_suspended": true,
                                        "status_reason": "Tindakan moderasi oleh Super Admin"
                                    },
                                    "errors": null
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/companies": {
            "get": {
                "tags": [
                    "Super Admin Moderation"
                ],
                "summary": "[Super Admin] Daftar Perusahaan Terdaftar",
                "description": "Mengambil daftar profil seluruh perusahaan mitra.",
                "operationId": "efd6352435e0bac9c62c58a40fabf8c8",
                "responses": {
                    "200": {
                        "description": "Daftar profil perusahaan",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Daftar perusahaan berhasil diambil.",
                                    "data": {
                                        "current_page": 1,
                                        "data": [
                                            {
                                                "id": 1,
                                                "company_name": "PT Tech Nusantara",
                                                "is_verified": true
                                            }
                                        ]
                                    },
                                    "errors": null
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/companies/{id}/toggle-verify": {
            "patch": {
                "tags": [
                    "Super Admin Moderation"
                ],
                "summary": "[Super Admin] Verifikasi / Unverify Perusahaan",
                "description": "Memberikan centang biru verifikasi resmi kepada profil perusahaan.",
                "operationId": "1c0cee98a42b471e2a10960f50f42fa6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status verifikasi diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Status verifikasi perusahaan PT Tech Nusantara berhasil diubah menjadi terverifikasi.",
                                    "data": {
                                        "id": 1,
                                        "company_name": "PT Tech Nusantara",
                                        "is_verified": true
                                    },
                                    "errors": null
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/role-requests": {
            "get": {
                "tags": [
                    "Super Admin Moderation"
                ],
                "summary": "[Super Admin] Daftar Pengajuan Role HR/Company",
                "description": "Mengambil permohonan penambahan role HR / Company Owner.",
                "operationId": "a51f259d9b907a23be0d76c68fcab68e",
                "responses": {
                    "200": {
                        "description": "Daftar pengajuan role",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Daftar pengajuan role berhasil diambil.",
                                    "data": {
                                        "current_page": 1,
                                        "data": [
                                            {
                                                "id": 1,
                                                "user_id": 1,
                                                "requested_role": "HR",
                                                "company_name": "PT Tech Nusantara",
                                                "status": "pending"
                                            }
                                        ]
                                    },
                                    "errors": null
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/role-requests/{id}/approve": {
            "post": {
                "tags": [
                    "Super Admin Moderation"
                ],
                "summary": "[Super Admin] Setujui Permohonan Role HR",
                "description": "Setujui permohonan role HR/Company Owner dan tetapkan role Spatie kepada user.",
                "operationId": "f68bf6897a22b2d687acd35ff9af842b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role request disetujui",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Permohonan role HR untuk Budi Pratama disetujui!",
                                    "data": {
                                        "id": 1,
                                        "requested_role": "HR",
                                        "status": "approved"
                                    },
                                    "errors": null
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/audit-logs": {
            "get": {
                "tags": [
                    "Super Admin Moderation"
                ],
                "summary": "[Super Admin] Audit Logs Keamanan Platform",
                "description": "Mengambil catatan riwayat log aktivitas keamanan pengguna.",
                "operationId": "072b2439383c5934cc39ecf90aecdc43",
                "responses": {
                    "200": {
                        "description": "Audit logs",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Log audit sistem berhasil diambil.",
                                    "data": {
                                        "current_page": 1,
                                        "data": [
                                            {
                                                "id": 1,
                                                "user_id": 1,
                                                "action": "LOGIN",
                                                "ip_address": "127.0.0.1",
                                                "created_at": "2026-08-21T22:00:00.000000Z"
                                            }
                                        ]
                                    },
                                    "errors": null
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Sanctum Bearer Token to access protected API endpoints.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}